Move ActivityTable to CommonTables (new root folder)
