+++ src/components/CommonTables/ActivityTable/helpers.ts
