+++ src/components/CommonTables/ActivityTable/hooks.ts
@@ -6,9 +6,9 @@ import { useInfiniteQuery } from '@tanstack/react-query'
 import { RBOption } from '@banx/components/RadioButton'
 
 import { fetchLenderActivity } from '@banx/api/activity'
+import { useMarketsPreview } from '@banx/pages/LendPage'
 import { useTokenType } from '@banx/store'
 
-import { useMarketsPreview } from '../../hooks'
 import { ActivityEvent, RADIO_BUTTONS_OPTIONS } from './constants'
 import { appendIdToOptions } from './helpers'
