+++ src/components/CommonTables/ActivityTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ActivityTable'
